部分修改

+++ utils/exists_mkdir.py
@@ -17,9 +17,9 @@ class ExistsMkDir:
         self._original_data = f"{self._path}/compare/original_data/{self.parsing}/{time.strftime('%Y-%m-%d', time.localtime())}/"
         self.path_lists = [self._excel_report, self._result_file, self._challenger, self._original_data]
 
-    def exists_mk_dir(self):
+    def exists_mk_dir(self, bol=True):
         for x in self.path_lists:
-            if x in (f"{self._result_file}", f"{self._excel_report}"):
+            if x in (f"{self._result_file}", f"{self._excel_report}") and bol:
                 if Path(x).exists():
                     shutil.rmtree(x, ignore_errors=True)
                 Path(x).mkdir()
